Move contrib matmul operators to a separate file
Move the `contrib` submodule of `src/ops/matmul.rs` to
`src/ops/matmul/contrib.rs` and move the tests for the contrib
operators into a `tests` module in that file.
